feat: add export script
This script is used to properly export any required classes to the browser.

When using `chrome.scripting.executeScript` the function body of the passed function is executed almost literally and thus does not have access to any classes or other objects defined in the parent scope.

We mitigate this by first injecting this `exports.ts` file, which will then prepare the global scope and make any required classes or other objects available.